7311: Refactor PeerManager to be an interface
